+++ packages/eslint-plugin-template/src/rules/click-events-have-key-events.ts
@@ -3,12 +3,21 @@ import { getTemplateParserServices } from '@angular-eslint/utils';
 import { createESLintRule } from '../utils/create-eslint-rule';
 import { getDomElements } from '../utils/get-dom-elements';
 import { isHiddenFromScreenReader } from '../utils/is-hidden-from-screen-reader';
-import { isInteractiveElement } from '../utils/is-interactive-element';
+import { isInherentlyInteractiveElement } from '../utils/is-interactive-element';
 import { isPresentationRole } from '../utils/is-presentation-role';
+import { getAttributeValue } from '../utils/get-attribute-value';
+import { ARIARole } from 'aria-query';
 
-export type Options = [];
+export type Options = [
+  {
+    readonly ignoreWithDirectives?: string[];
+  },
+];
 export type MessageIds = 'clickEventsHaveKeyEvents';
 export const RULE_NAME = 'click-events-have-key-events';
+const DEFAULT_OPTIONS: Options[number] = {
+  ignoreWithDirectives: [],
+};
 
 export default createESLintRule<Options, MessageIds>({
   name: RULE_NAME,
@@ -18,14 +27,29 @@ export default createESLintRule<Options, MessageIds>({
       description:
         '[Accessibility] Ensures that the click event is accompanied with at least one key event keyup, keydown or keypress.',
     },
-    schema: [],
+    schema: [
+      {
+        type: 'object',
+        properties: {
+          ignoreWithDirectives: {
+            type: 'array',
+            items: { type: 'string' },
+            uniqueItems: true,
+            default: DEFAULT_OPTIONS.ignoreWithDirectives as
+              | string[]
+              | undefined,
+          },
+        },
+        additionalProperties: false,
+      },
+    ],
     messages: {
       clickEventsHaveKeyEvents:
         'click must be accompanied by either keyup, keydown or keypress event for accessibility.',
     },
   },
-  defaultOptions: [],
-  create(context) {
+  defaultOptions: [DEFAULT_OPTIONS],
+  create(context, [{ ignoreWithDirectives }]) {
     return {
       Element(node: TmplAstElement) {
         if (!getDomElements().has(node.name)) {
@@ -33,10 +57,18 @@ export default createESLintRule<Options, MessageIds>({
         }
 
         if (
+          isIgnored(ignoreWithDirectives, node) ||
           isPresentationRole(node) ||
           isHiddenFromScreenReader(node) ||
-          isInteractiveElement(node)
+          isInherentlyInteractiveElement(node)
         ) {
+          return;
+        }
+
+        // The final case that should be ignored is element which is not inherently interactive, but which has an interactive role.
+        // TODO: extend utils with this check (and make it include all interactive roles)
+        const role = getAttributeValue(node, 'role') as ARIARole;
+        if (role === 'button') {
           return;
         }
 
@@ -67,3 +99,23 @@ export default createESLintRule<Options, MessageIds>({
     };
   },
 });
+
+function isIgnored(
+  ignoreWithDirectives: string[] | undefined,
+  { inputs, attributes }: TmplAstElement,
+) {
+  if (ignoreWithDirectives && ignoreWithDirectives.length > 0) {
+    for (const input of inputs) {
+      if (ignoreWithDirectives.includes(input.name)) {
+        return true;
+      }
+    }
+    for (const attribute of attributes) {
+      if (ignoreWithDirectives.includes(attribute.name)) {
+        return true;
+      }
+    }
+  }
+
+  return false;
+}