+++ packages/eslint-plugin-template/tests/rules/click-events-have-key-events/cases.ts
@@ -55,7 +55,18 @@ export const valid: readonly (string | ValidTestCase<Options>)[] = [
         <a href="#" (click)="onClick()"></a>
         <a [attr.href]="href" class="anchor" (click)="onClick()"></a>
         <a [routerLink]="'route'" (click)="onClick()"></a>
+        <div role="button" (click)="doSomething()"></div>
+        <span role="button" (click)="doSomething()"></span>
+        <p role="button" (click)="doSomething()"></p>
       `,
+  },
+  {
+    code: `<div myDirective (click)="onClick()"></div>`,
+    options: [{ ignoreWithDirectives: ['myDirective'] }],
+  },
+  {
+    code: `<div [myDirective] (click)="onClick()"></div>`,
+    options: [{ ignoreWithDirectives: ['myDirective'] }],
   },
 ];
 
@@ -157,5 +168,15 @@ export const invalid: readonly InvalidTestCase<MessageIds, Options>[] = [
       <div (click)="onClick()" [attr.hidden]="hidden"></div>
       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     `,
+  }),
+  convertAnnotatedSourceToFailureCase({
+    messageId,
+    description:
+      'should fail if an element has no key events, and ignoreWithDirectives option specifies directives, but none of the directives are present',
+    annotatedSource: `
+      <div myDirective (click)="onClick()"></div>
+      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    `,
+    options: [{ ignoreWithDirectives: ['testDirective', 'otherDirective'] }],
   }),
 ];