+++ packages/eslint-plugin-template/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 19.6.0 (2025-05-27)
+
+### 🩹 Fixes
+
+- respect existing eslint.config.ts, eslint.config.cts, eslint.config.mts files ([#2458](https://github.com/angular-eslint/angular-eslint/pull/2458))
+
+### ❤️ Thank You
+
+- James Henry @JamesHenry
+
 ## 19.5.0 (2025-05-25)
 
 ### 🚀 Features